add warning for ruby with ghidra 10.0.3 (#10)
This patch updates the version of JRuby, and adds a warning message and
patch file to the plugin to address potential errors when used with
Ghidra 10.0.3. These changes address an issue regarding class lookups
from plugins in newer JDK versions arising from illegal reflective
accesses from the Apache Felix framework.

This addresses #9, which can be referenced for more details.