update for ghidra 10.2 (#27)
Updates JRuby, Kotlin, and the CI builds for Ghidra 10.2 support.

This is a major release as the getScriptInstance method signatures changed
in a way that may break compilation of other programs relying on them.
Otherwise, functionality is unchanged.

The `test` target has been disabled in CI as it gives inconsistent and
misleading failures. These need to re-evaluated to see if there is a more
portable and reliable way for them to work.